adopted test, CI

+++ tests/protzilla/test_run.py
@@ -1,3 +1,4 @@
+from os import path
 from shutil import rmtree
 
 from protzilla import data_preprocessing
@@ -8,7 +9,8 @@ from protzilla.run import Run
 
 def test_run_create():
     # here the run should be used like in the CLI
-    rmtree(PATH_TO_RUNS / "test_run")
+    if path.exists(PATH_TO_RUNS / "test_run"):
+        rmtree(PATH_TO_RUNS / "test_run")
     run = Run.create("test_run")
     run.calculate_and_next(
         main_data_import.max_quant_import,
